+++ LambdaEngine/Include/Audio/FMOD/AudioDeviceFMOD.h
@@ -29,9 +29,9 @@ namespace LambdaEngine
 		virtual IAudioGeometry*		CreateAudioGeometry(const AudioGeometryDesc* pDesc)		override final;
 		virtual IReverbSphere*		CreateReverbSphere(const ReverbSphereDesc* pDesc)		override final;
 
-		virtual void SetMasterVolume(float volume) override final;
+		virtual void SetMasterVolume(float64 volume) override final;
 
-		virtual float GetMasterVolume() const override final;
+		virtual float64 GetMasterVolume() const override final;
 
 	public:
 		FMOD_SYSTEM* pSystem					= nullptr;
