+++ LambdaEngine/Include/Audio/Lambda/ManagedSoundInstance3DLambda.h
@@ -23,7 +23,7 @@ namespace LambdaEngine
 		uint32				Flags				= FSoundModeFlags::SOUND_MODE_NONE;
 		glm::vec3			Position			= glm::vec3(0.0f);
 		glm::vec3			Velocity			= glm::vec3(0.0f);
-		float32				Volume				= 1.0f;
+		float64				Volume				= 1.0;
 		float32				Pitch				= 1.0f;
 	};
 
@@ -35,7 +35,7 @@ namespace LambdaEngine
 
 		bool Init(const ManagedSoundInstance3DDesc* pDesc);
 
-		void UpdateVolume(float masterVolume, const AudioListenerDesc* pAudioListeners, uint32 audioListenerCount, ESpeakerSetup speakerSetup);
+		void UpdateVolume(float64 masterVolume, const AudioListenerDesc* pAudioListeners, uint32 audioListenerCount, ESpeakerSetup speakerSetup);
 		void AddToBuffer(double** ppOutputChannels, uint32 channelCount, uint32 outputSampleCount);
 
 	private:
@@ -51,7 +51,7 @@ namespace LambdaEngine
 		glm::vec3	m_Position						= glm::vec3(0.0f);
 		//Velocity
 		//Pitch			
-		float		m_Volume						= 1.0f;
+		float64		m_Volume						= 1.0;
 
 		float64*	m_pOutputVolumes				= nullptr;
 	};
