+++ LambdaEngine/Include/Audio/Lambda/SoundEffect3DLambda.h
@@ -15,7 +15,7 @@ namespace LambdaEngine
 		~SoundEffect3DLambda();
 
 		virtual bool Init(const SoundEffect3DDesc* pDesc) override final;
-		virtual void PlayOnceAt(const glm::vec3& position, const glm::vec3& velocity, float volume, float pitch) override final;
+		virtual void PlayOnceAt(const glm::vec3& position, const glm::vec3& velocity, float64 volume, float pitch) override final;
 
 		FORCEINLINE virtual float64 GetDuration() override final { return (float64)m_Header.Duration; }
