+++ LambdaEngine/Include/Audio/Lambda/SoundInstance3DLambda.h
@@ -26,11 +26,11 @@ namespace LambdaEngine
 		virtual void Toggle()	override final;
 
 		virtual void SetPosition(const glm::vec3& position) override final;
-		virtual void SetVolume(float32 volume)				override final;
-		virtual void SetPitch(float32 pitch)					override final;
+		virtual void SetVolume(float64 volume)				override final;
+		virtual void SetPitch(float32 pitch)				override final;
 
 		FORCEINLINE virtual const glm::vec3&	GetPosition()	const override final { return m_Position;	}
-		FORCEINLINE virtual float32				GetVolume()		const override final { return m_Volume;		}
+		FORCEINLINE virtual float64				GetVolume()		const override final { return m_Volume;		}
 		FORCEINLINE virtual float32				GetPitch()		const override final { return 1.0f;			}
 
 		void UpdateVolume(float masterVolume, const AudioListenerDesc* pAudioListeners, uint32 audioListenerCount, ESpeakerSetup speakerSetup);
@@ -49,7 +49,7 @@ namespace LambdaEngine
 
 		glm::vec3	m_Position						= glm::vec3(0.0f);
 		//Pitch			
-		float32		m_Volume						= 1.0f;
+		float64		m_Volume						= 1.0;
 
 		float64*	m_pOutputVolumes				= nullptr;
 	};
